pyproject.toml
[tool.poetry]
name = "proyecto0"
version = "0.1.0"
description = "proyecto de nivelación de cloud"
authors = ["Camilo Rodriguez <juancamilorodriguezfonseca@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.8"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
black = "^25.1.0"
sqlalchemy = "^2.0.37"
uvicorn = "^0.34.0"
pyjwt = "^2.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
